Generator
Fibonacci series as we all know is a series of numbers wherein each number is a sum of preceding two numbers. The first two numbers are 1 and 1.

Memory is saved as the items are produced when required, unlike normal Python functions. This fact becomes very important when you need to create a huge number of iterators. This is also considered as the biggest advantage of generators.